requirements.txt
# Core AI/LLM
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5

# Data Processing
pandas>=2.0.0
duckdb>=0.9.0
pyarrow>=14.0.0

# Frontend
streamlit>=1.30.0
altair>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging
loguru>=0.7.0
